fixes some compile warnings
git-svn-id: http://svn.ratbox.org/svnroot/trunk@6168 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/linebuf.h
@@ -53,5 +53,5 @@ extern int linebuf_get(buf_head_t *, char *, int);
 extern void linebuf_put(buf_head_t *, char *, int);
 extern int linebuf_flush(int, buf_head_t *);
 
-extern void count_linebuf_memory( int *count, int *memory_used);
+extern void count_linebuf_memory( int *count, u_long *memory_used);
 #endif
